requirements.txt
asgiref==3.5.0
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.1
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-modeltranslation==0.17.5
environs==9.5.0
folium==0.12.1.post1
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
numpy==1.22.1
Pillow==9.0.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
